tabs ported to new api (thanks uzi) and compilation fix on 64bit systems
git-svn-id: svn+ssh://kadu.net/home/kadu/repos/kadu/trunk@5972 fea5205d-6df5-0310-bef7-ea67e47da9a5

+++ kadu-core/userbox.cpp
@@ -1656,7 +1656,7 @@ inline int compareStatus(const UserListElement &u1, const UserListElement &u2)
 	if (d1 && d2)
 		return d2->status().compareTo(d1->status());
 	else
-		return (int)d2 - (int)d1;
+		return (long int)d2 - (long int)d1;
 }
 
 void UserBox::accountRegistered(Account *account)
